+++ TWCManager.py
@@ -93,6 +93,7 @@ modules_available = [
     "EMS.HASS",
     "EMS.Kostal",
     "EMS.OpenHab",
+    "EMS.OpenWeatherMap",
     "EMS.SmartMe",
     "EMS.SmartPi",
     "EMS.SolarEdge",
@@ -342,7 +343,7 @@ def update_statuses():
         chgwattsDisplay = f("{chgwatts:.0f}W")
 
         if config["config"]["subtractChargerLoad"]:
-            if connwatts > 0:
+            if cnnwatts > 0:
                 othwatts = conwatts - chgwatts
             else:
                 othwatts = 0
@@ -378,7 +379,7 @@ def update_statuses():
 
         nominalOffer = master.convertWattsToAmps(
             genwatts + (chgwatts if (config["config"]["subtractChargerLoad"] and conwatts == 0) else 0)
-            - (conwatts - (chgwatts if (config["config"]["subtractChargerLoad"] and connwatts > 0) else 0))
+            - (conwatts - (chgwatts if (config["config"]["subtractChargerLoad"] and conwatts > 0) else 0))
         )
         if abs(maxamps - nominalOffer) > 0.005:
             nominalOfferDisplay = f("{nominalOffer:.2f}A")